third_party/WebKit/LayoutTests/crypto/subtle/ecdh/import-jwk-expected.txt

Test importing an EC public and private key for ECDH.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Importing a public key...
PASS publicKey.toString() is '[object CryptoKey]'
PASS publicKey.type is 'public'
PASS publicKey.usages is []
PASS publicKey.algorithm.name is 'ECDH'
PASS publicKey.algorithm.namedCurve is 'P-256'

Importing a private key...
PASS privateKey.toString() is '[object CryptoKey]'
PASS privateKey.type is 'private'
PASS privateKey.usages is ['deriveBits']
PASS privateKey.algorithm.name is 'ECDH'
PASS privateKey.algorithm.namedCurve is 'P-384'
PASS successfullyParsed is true

TEST COMPLETE
